backend/pyproject.toml
[project]
name = "ojas-backend"
version = "0.1.0"
description = "OJAS Ayurvedic App - FastAPI Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "firebase-admin>=6.5.0",
    "motor>=3.6.0",
    "pymongo>=4.10.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "itsdangerous>=2.2.0",
    "httpx>=0.27.0",
    "email-validator>=2.3.0",
]
